requirements.txt
numpy>=1.11.0
INDI-Tools>=0.0.6
Pillow>=3.2.0
PyPDF2>=1.26.0
PyYAML>=3.11
argparse>=1.2.1
boto3>=1.3.1
botocore>=1.4.22
configparser>=3.5.0
cycler>=0.10.0
decorator>=4.0.9
docutils>=0.12
future>=0.15.2
futures>=3.0.5
html5lib>=1.0b8
httplib2>=0.9.2
jmespath>=0.9.0
matplotlib>=1.5.1
networkx>=1.11
nibabel>=2.0.2
nitime>=0.6
nipype>=0.12.1
nose>=1.3.7
pandas>=0.18.1
prov>=1.4.0
pyparsing>=2.1.4
python-dateutil>=2.5.3
pytz>=2016.4
reportlab>=3.3.0
scipy>=0.17.1
seaborn>=0.7.0
simplejson>=3.8.2
six>=1.10.0
traits>=4.5.0
wsgiref>=0.1.2
xhtml2pdf>=0.1a4
